split to error_handlers.rb
